refactor(trpc middleware): refacto trpc middleware
