+++ apps/recnet/src/server/routers/middleware.ts
@@ -7,6 +7,69 @@ import {
 } from "@recnet/recnet-jwt";
 import { UserRole } from "@recnet/recnet-web/constant";
 import { z } from "zod";
+import { Tokens } from "next-firebase-auth-edge";
+import { userSchema, userPreviewSchema } from "@recnet/recnet-api-model";
+import { db } from "@recnet/recnet-web/firebase/admin";
+import { notEmpty } from "@recnet/recnet-web/utils/notEmpty";
+
+/**
+ * @param tokens Tokens: user tokens from next-firebase-auth-edge
+ * @returns User: parsed by userSchema
+ *
+ * Note: Internal function and used in trpc middlewares
+ */
+async function _getUserByTokens(tokens: Tokens) {
+  // REFACTOR_AFTER_MIGRATION
+  const { decodedToken } = tokens;
+  const email = decodedToken.email as string;
+
+  const querySnapshot = await db
+    .collection("users")
+    .where("email", "==", email)
+    .limit(1)
+    .get();
+  if (querySnapshot.empty) {
+    throw new TRPCError({
+      code: "NOT_FOUND",
+      message: "User not found",
+    });
+  }
+  // get userPreview for each following
+  const userPreviews = await Promise.all(
+    querySnapshot.docs[0].data().following.map(async (id: string) => {
+      const unparsedUser = await db.doc(`users/${id}`).get();
+      const userData = unparsedUser.data();
+      if (!userData) {
+        return null;
+      }
+      const parsedRes = userPreviewSchema.safeParse({
+        id: unparsedUser.id,
+        handle: userData.username,
+        displayName: userData.displayName,
+        photoUrl: userData.photoURL,
+        affiliation: userData.affiliation || null,
+        bio: userData.bio || null,
+        numFollowers: userData.followers.length,
+      });
+      if (parsedRes.success) {
+        return parsedRes.data;
+      }
+      return null;
+    })
+  );
+  return userSchema.parse({
+    id: querySnapshot.docs[0].id,
+    handle: querySnapshot.docs[0].data().username,
+    displayName: querySnapshot.docs[0].data().displayName,
+    photoUrl: querySnapshot.docs[0].data().photoURL,
+    affiliation: querySnapshot.docs[0].data().affiliation || null,
+    bio: querySnapshot.docs[0].data().bio || null,
+    numFollowers: querySnapshot.docs[0].data().followers.length,
+    email: querySnapshot.docs[0].data().email,
+    role: querySnapshot.docs[0].data().role ? UserRole.ADMIN : UserRole.USER,
+    following: userPreviews.filter(notEmpty),
+  });
+}
 
 export const checkFirebaseJWTProcedure = publicProcedure.use(async (opts) => {
   const tokens = await getTokenServerSide();
@@ -37,9 +100,11 @@ export const checkRecnetJWTProcedure = publicProcedure.use(async (opts) => {
       message: "Unauthorized, missing Recnet secret",
     });
   }
+  const user = await _getUserByTokens(tokens);
   return opts.next({
     ctx: {
       tokens: parseRes.data,
+      user: user,
     },
   });
 });
@@ -63,9 +128,11 @@ export const checkIsAdminProcedure = publicProcedure.use(async (opts) => {
       });
     }
   }
+  const user = await _getUserByTokens(tokens);
   return opts.next({
     ctx: {
-      tokens,
+      tokens: parseRes.data,
+      user,
     },
   });
 });
